check the course existing (#168)
Consultation Booking Hotfix

+++ backend/utility_module/ConsultationManager.py
@@ -6,6 +6,7 @@ from database.DataBaseManager import DataBaseManager
 from utility_module.emailer.EmailSender import EmailSender
 import datetime
 from conf.Logger import Logger
+from conf.Error import ConsultationError
 
 logger = Logger(__name__).log
 
@@ -43,12 +44,14 @@ class ConsultationManager:
         return self.database_manager.execute_query(query, inputs)
 
     def delete_consultation(self, cid, sid, time, date):
+        if not self.check_course_exist(cid):
+           return ConsultationError.INVALID_COURSE.value
         check_empty = "Select * FROM consultation WHERE cid = %s and sid = %s and time = %s and date = %s"
         query = "DELETE FROM consultation WHERE cid = %s and sid = %s and time = %s and date = %s"
         inputs = (cid.upper(), sid, time, date)
         data_exist = self.database_manager.execute_query(check_empty, inputs)
         if not data_exist:
-            return "There is no course consultation booked at this time"
+            return ConsultationError.INVALID_TIME.value
         self.emailer.send_confirm_cancelling(cid=cid, time=time, date=date, receiver='whatbot9900@gmail.com')
         return self.database_manager.execute_query(query, inputs)
 
@@ -167,6 +170,8 @@ class ConsultationManager:
         :return: success status
         :rtype: str
         """
+        if not self.check_course_exist(cid):
+            return ConsultationError.INVALID_COURSE.value
         is_weekday, feedback = self.check_weekday(date)
         time = self.round_time(time)
         if is_weekday:
@@ -223,4 +228,20 @@ class ConsultationManager:
         """
         hour, mins, _ = time.split(":")
         return '{:02d}:00:00'.format(int(hour)+1 ) if int(mins) >= 30 else '{:02d}:00:00'.format(int(hour))
+
+    def check_course_exist(self, cid):
+       course_codes = ['COMP9900', 'comp9321', 'COMP9945', 'COMP9101', 'COMP9041', 'COMP9331', 'COMP9311',
+                      'COMP9414', 'COMP9841', 'COMP6451', 'COMP9024', 'COMP9021', 'COMP9020', 'COMP9322',
+                      'COMP6714', 'COMP6771', 'COMP9153', 'COMP9313', 'COmp9417', 'COMP9444', 'COMP9334',
+                      'COMP9517', 'COMP9201', 'COMP9102', 'COMP9315', 'COMP4121', 'COMP9323', 'COMP9318',
+                      'COMP6441', 'comp9511', 'ComP9032', 'Comp4418', 'comP6324', 'CoMp9415', 'ComP4141',
+                      'COmP6752', 'comP9211', 'comP9319', 'cOMP9336', 'comP6471', 'COMP9243', 'COMP9283',
+                      'COMP5752', 'comp9814', 'GSOE9210', 'GSOE9220', 'GSOE9820', 'COMP9801', 'COMP9222',
+                      'COMP9447', 'COMP4161', 'COMP6452', 'COMP6733', 'COMP6841', 'COMP9151', 'COMP9161',
+                      'COMP9332', 'COMP9333', 'COMP9337', 'COMP9431', 'COMP6443', 'COMP6445', 'COMP6447',
+                      'COMP6448', 'COMP6741', 'COMP6843', 'COMP6845', 'COMP9242', 'COMP9418', 'COMP9596',
+                       'COMP6721', 'COMP9154', 'COMP9164', 'COMP9434', 'COMP9424', 'COMP9423']
+       if(cid.upper() in course_codes):
+           return True
+       return False
